removed more unneeded impls

+++ malachite-test/src/integer/arithmetic/eq_limb_mod_limb.rs
@@ -2,6 +2,7 @@ use malachite_base::num::arithmetic::traits::{DivisibleBy, EqMod, Mod};
 use malachite_base::num::conversion::traits::CheckedFrom;
 use malachite_base::num::logic::traits::SignificantBits;
 use malachite_nz::integer::arithmetic::eq_limb_mod_limb::limbs_eq_neg_limb_mod_limb;
+use malachite_nz::integer::Integer;
 use malachite_nz::platform::Limb;
 
 use common::{m_run_benchmark, BenchmarkType, DemoBenchRegistry, GenerationMode, ScaleType};
@@ -160,7 +161,7 @@ fn benchmark_integer_eq_limb_mod_limb_algorithms(
             ),
             (
                 "(Integer - Limb).divisible_by(Limb)",
-                &mut (|(n, u, modulus)| no_out!((n - u).divisible_by(modulus))),
+                &mut (|(n, u, modulus)| no_out!((n - Integer::from(u)).divisible_by(modulus))),
             ),
         ],
     );
